Refactor: group storefront home components into a folder
