+++ src/components/storefront/index.js
@@ -1,6 +1,6 @@
 import Navbar from './Navbar/Navbar';
-import BestSellers from './BestSellers/BestSellers';
-import ItemListings from './ItemListings/ItemListings';
+import BestSellers from './Home/BestSellers/BestSellers';
+import AllItems from './Home/AllItems/AllItems';
 import Footer from './Footer/Footer';
 import TeamBanner from './TeamBanner/TeamBanner';
 import StorefrontRouter from './StorefrontRouter/StorefrontRouter';
@@ -11,7 +11,7 @@ import CartItem from './CartItem/CartItem';
 export {
   Navbar,
   BestSellers,
-  ItemListings,
+  AllItems,
   Footer,
   TeamBanner,
   AdditionalContent,
